fix: CMD+F not working on screens with keyboard shortcut guide
